Lab0/requirements.txt
numpy~=1.20.1
matplotlib~=3.3.4
sounddevice~=0.4.1
soundfile~=0.10.3.post1
opencv-python~=4.5.1.48
future~=0.18.2
pillow~=8.1.0
numba~=0.52.0
scipy~=1.6.1
pympler~=0.9
tqdm~=4.59.0